requirements.txt
absl-py==2.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.0
easydict==1.10
fonttools==4.43.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lml==0.1.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
pyexcel-ods3==0.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==1.12.1
torch-geometric==2.2.0
torch-scatter==2.1.1
torch-sparse==0.6.17
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
